requirements.txt
crewai
python-dotenv
pandas
numpy
matplotlib
seaborn
reportlab
pydantic
google-generativeai  # For Gemini API 